+++ ArcdpsLogManager/CHANGELOG.md
@@ -2,10 +2,11 @@
 
 This is the full changelog of the arcdps Log Manager.
 
-## Log Manager v1.10.1
+## Log Manager v1.11
 
 #### New features
 - Added Cosmic Observatory CM detection
+- Added Temple of Febe CM detection
 
 #### Fixes
 - Fixed a very rare Harvest Temple false failure